feat: 식당 상세페이지 domain, data layer 추가 및 UI 수정

+++ domain/src/main/java/com/example/domain/model/map/VeganMapRestaurant.kt
@@ -1,7 +1,8 @@
 package com.example.domain.model.map
 
+
 data class VeganMapRestaurant(
-    val id: Int,
+    val id: Long,
     val name: String,
     val type: String?,
     val distance: Double,
